@@ -19,28 +19,28 @@
     "test:unit": "vitest"
   },
   "devDependencies": {
-    "@playwright/test": "^1.44.1",
-    "@sveltejs/adapter-cloudflare": "^4.4.0",
-    "@sveltejs/kit": "^2.5.10",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
-    "@types/eslint": "^8.56.10",
-    "@typescript-eslint/eslint-plugin": "^7.12.0",
-    "@typescript-eslint/parser": "^7.12.0",
-    "eslint": "^8.57.0",
+    "@playwright/test": "^1.48.2",
+    "@sveltejs/adapter-cloudflare": "^4.7.4",
+    "@sveltejs/kit": "^2.7.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@types/eslint": "^9.6.1",
+    "@typescript-eslint/eslint-plugin": "^8.12.2",
+    "@typescript-eslint/parser": "^8.12.2",
+    "eslint": "^9.13.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.39.0",
-    "prettier": "^3.3.1",
-    "prettier-plugin-svelte": "^3.2.4",
-    "svelte": "^4.2.18",
-    "svelte-check": "^3.8.0",
-    "tslib": "^2.6.3",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.12",
-    "vitest": "^1.6.0"
+    "eslint-plugin-svelte": "^2.46.0",
+    "prettier": "^3.3.3",
+    "prettier-plugin-svelte": "^3.2.7",
+    "svelte": "^5.1.9",
+    "svelte-check": "^4.0.5",
+    "tslib": "^2.8.1",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10",
+    "vitest": "^2.1.4"
   },
   "dependencies": {
     "d3-dsv": "^3.0.1",
-    "maplibre-gl": "^4.3.2",
+    "maplibre-gl": "^5.0.0-pre.4",
     "normalize.css": "^8.0.1"
   }
 }
