@@ -8,7 +8,7 @@
   import { onMount } from 'svelte';
 
   const { Map, NavigationControl, GeolocateControl, setRTLTextPlugin } = MapLibreGL;
-  let mapContainer: HTMLDivElement;
+  let mapContainer: HTMLDivElement = $state();
 
   onMount(async () => {
     setRTLTextPlugin('/scripts/maplibre-gl-rtl-text.min.js', true);
@@ -34,7 +34,7 @@
   });
 </script>
 
-<div class="map" id="map" bind:this={mapContainer} />
+<div class="map" id="map" bind:this={mapContainer}></div>
 
 <style>
   .map {
